ASCA_permutation

the function implements a design permutation strategy to validate the results of an ASCA decomposition

ASCA_permutation(asca, turns = 100, qt = 0.95)

Arguments

asca

an object of class ASCA_decomposition

turns

the number of permutations

qt

the empirical quantile

Value

A list with the following elements

R2_qt: a matrix with the quantiles for the pseusoR2 of the individual variables

L2_qt: a matrix with the quantiles for the L2 norm of the individual terms

varimp_qt: a list holding the matrices of quantiles for each variable in each term

Examples

## load the data
data("synth_count_data")

## perform the ASCA decomposition
dec_test <- ASCA_decompose( d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)

## validate the outcomes

val_out <- ASCA_permutation(dec_test, 10)
#> Permutation # 1 
#> Permutation # 2 
#> Permutation # 3 
#> Permutation # 4 
#> Permutation # 5 
#> Permutation # 6 
#> Permutation # 7 
#> Permutation # 8 
#> Permutation # 9 
#> Permutation # 10